+++ lib/dsa_kit.dart
@@ -1,3 +1,6 @@
 export 'heap/heap_base.dart';
 export 'heap/min_heap.dart';
 export 'heap/max_heap.dart';
+
+export 'dequeue/dequeue_base.dart';
+export 'dequeue/deque.dart';